requirements.txt
pandas
numpy
tqdm
matplotlib
mplfinance
pymysql
psycopg2-binary
duckdb
requests
tushare
akshare
pytdx
mootdx
fastapi
uvicorn[standard]
pydantic
starlette
pytest
